+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to
 ### Changed
 
 - This library has been renamed to `liberator.resource.discovery`.
+- Rather than `handler` accepting an `options` map, it now accepts an
+  `overrides` map, used as the most specific definition map when building the
+  resource. As such, anything that would normally go into a liberator resource
+  definition map can be provided.
+- Rather than providing a `:links` option with details of the links to add to
+  the discovery response, now a `:link-definitions` key should be provided in
+  the `overrides` map, whose value can be:
+    - a map of link names to link parameters
+    - a vector of keyword route names
+    - a vector of link parameters
+    - a vector of functions of context producing link maps
+- The default links have now been removed.
 
 ## 0.1.0 — 2020-03-30